Split Residue.jl.
